% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cols_var.R
\name{col_var}
\alias{col_var}
\title{Variance by columns}
\usage{
col_var(x)
}
\arguments{
\item{x}{A matrix or dataframe.}
}
\value{
The variance values by column from dataset \code{x}
}
\description{
Variance by columns
}
\examples{
cols_var(iris)

}
