% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cols_NA.R
\name{cols_NAs}
\alias{cols_NAs}
\title{Counts NA values by column}
\usage{
cols_NAs(x)
}
\arguments{
\item{x}{A matrix or dataframe.}
}
\value{
The NA values by column from dataset \code{x}
}
\description{
Counts NA values by column
}
\examples{
cols_NAs(iris)

}
